### 2024-09-26 #7283 @iosefa
> Hi @iosefa and thanks for your submission! I am looking for some specific items to make sure your submission fits our requirements at a high level (not at the more detailed review level) before moving on to finding an editor or putting this on our waitlist if no relevant editors are available. I'll comment over time as I have a chance to go through them:
>
> * [ ]  [OSI-approved](https://opensource.org/licenses) license
> * [ ]   Installation instructions
> * [ ]   Documentation exists at all
> * [ ]  Tests of some sort
> * [ ]  Check paper (make sure compiles, read through, check for length [250-1000 words](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain))
> * [ ]   [Substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
> * [ ]  Clear research application
>
> In the meantime, please take a look at the comments above ⬆️ from the editorialbot to address any DOI, license, or paper issues if you're able (there may not be any), or suggest reviewers. For reviewers, please suggest 5 reviewers from the [database listed above](https://reviewers.joss.theoj.org/lookup) or your own ([non-conflicted](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)) extended network. Their github handles are most useful to receive but please don't use "@" to reference them since it will prematurely ping them.

Hi @kthyng,
Thank you! I will go through the reviewer list and reply with suggestions. In the meantime, I have two questions:

1. I need to add an additional statement in the acknowledgements -- is it okay to do this at this point?

2. the editorial bot came back with this missing DOI:

❌ MISSING DOIs

- 10.32614/cran.package.leafr may be a valid DOI for title: leafR: Calculates the leaf area index (LAD) and ot...

I dont know that this is a correct and valid DOI for the leafR package. The package itself requests that it be cited as:

```R
citation("leafR")
To cite package ‘leafR’ in publications use:

  Almeida D, Stark S, Silva C, Hamamura C, Valbuena R (2021).
  _leafR: Calculates the Leaf Area Index (LAD) and Other Related
  Functions_. R package version 0.3.5,
  <https://CRAN.R-project.org/package=leafR>.

A BibTeX entry for LaTeX users is

  @Manual{,
    title = {leafR: Calculates the Leaf Area Index (LAD) and Other Related Functions},
    author = {Danilo Roberti Alves de Almeida and Scott Christopher Stark and Carlos Alberto Silva and Caio Hamamura and Ruben Valbuena},
    year = {2021},
    note = {R package version 0.3.5},
    url = {https://CRAN.R-project.org/package=leafR},
  }
```

And this is how it is cited in our manuscript. Is it okay to leave it as is without any specific reply to the editor bot?

Many thanks,
