### 2024-09-26 60eb564
Add EPT reader support to read_lidar function
Extended the read_lidar function to support 'ept.json' file types by recognizing the EPT reader. Also included a new 'bounds' parameter for spatial cropping and streamlined the pipeline building process for better maintenance. Updated tests to match the new validation error message.
