### 2024-09-26 7f2226b
Support WKT input for polygon geometry in crop function
Allow the `poly` parameter to accept either a file path or directly a WKT string for defining the polygon geometry used in cropping. This enhancement simplifies user input by recognizing and processing WKT strings directly.
