### 2024-09-26 0c2d65e
Update FHD calculation and documentation
Change the term from Frequency Histogram Diversity to Foliage Height Diversity for clarity. Enhanced function documentation to better explain the input parameters and the output format while highlighting areas with no voxel returns resulting in NaN values.
