### 2024-09-26 91cfd0e
Refactor polygon cropping validation
Simplified the validation logic for polygon input when cropping to a polygon. Now, the code clearly separates the responsibilities of checking for provided polygons and verifying file existence, improving readability and maintainability.
