### 2024-09-26 13a1eef
Add minor_version and in_srs parameters to pipeline steps
These changes ensure that the reprojection filter correctly identifies the input SRS with the `in_srs` parameter. Additionally, specifying `minor_version` in the output step aligns the format with expected versioning.
