### 2024-09-26 #7283 @kthyng
> I need to add an additional statement in the acknowledgements -- is it okay to do this at this point?

Yes go ahead.

> the editorial bot came back with this missing DOI:

Sometimes the DOIs are incorrect. If you checked and know better, that is fine. You'll want to let your topic editor know this later, too.
