### 2024-09-26 #7283 @kthyng
@iosefa
* I'm not in your field, but it seems like some plots to help users understand your docs would be great — any possibility of that? These can also act as a tutorial.
* your parenthetical references shouldn't have extra parentheses between them — check out the sample paper for exact syntax so they end up in a list together.
* your inline references shouldn't have parentheses
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, "r" is not capitalized.
