### 2024-09-26 #7283 @iosefa
@kthyng
Thank you very much for your feedback.
- Yes, I can and will add figures to the docs.
- will fix
- will fix
- will check and fix

I will work these in. I assume it is okay to commit directly to the branch with the paper?
