### 2024-09-29 46a2611
Fix format for references: - Add colons between author names and the year of publication in inline references. - Add semicolons between references with multiple publications - Update capitalization in .bib file so that titles are in title case, proper nouns are protected.
