+++ paper/paper.md
@@ -36,9 +36,9 @@ PyForestScan is an open-source Python library designed for calculating forest st
 
 # Statement of Need
 
-Remote sensing data, particularly Light Detection and Ranging (lidar) data from airborne sensors, are becoming increasingly accessible, offering a detailed understanding of forest ecosystems at fine spatial resolutions. This data is critical for calculating forest structural metrics such as canopy height, canopy cover, Plant Area Index (PAI), Plant Area Density (PAD), Foliage Height Diversity (FHD), as well as digital terrain models (DTM), which are essential for forest management, biodiversity conservation, and carbon accounting [@mcelhinnyForestWoodlandStand2005], [@drakeEstimationTropicalForest2002]. However, working with large-scale lidar datasets remains a challenge due to the complexity and size of the data. Additionally, despite Python being a powerful language widely used for geospatial and ecological analysis, there is a notable lack of dedicated, open-source tools within the Python ecosystem specifically designed for calculating these forest structural metrics from lidar data. Calculating these metrics is non-trivial, and several steps are often required to process the point clouds in order to generate these metrics. Existing open source solutions are primarily in the R programming language [@rousselLidRPackageAnalysis2020a] [@rousselAirborneLiDARData2024], [@dealmeidaLeafRCalculatesLeaf2021] or are proprietary, computationally intensive, or not flexible enough for the variety of ecological contexts in which these metrics are needed [@lastools], [@fusion], [@globalmapper]. This gap makes it difficult for researchers, ecologists, and forest managers to integrate lidar-based analysis into their workflows efficiently.
+Remote sensing data, particularly Light Detection and Ranging (lidar) data from airborne sensors, are becoming increasingly accessible, offering a detailed understanding of forest ecosystems at fine spatial resolutions. This data is critical for calculating forest structural metrics such as canopy height, canopy cover, Plant Area Index (PAI), Plant Area Density (PAD), Foliage Height Diversity (FHD), as well as digital terrain models (DTM), which are essential for forest management, biodiversity conservation, and carbon accounting [@mcelhinnyForestWoodlandStand:2005; @drakeEstimationTropicalForest:2002]. However, working with large-scale lidar datasets remains a challenge due to the complexity and size of the data. Additionally, despite Python being a powerful language widely used for geospatial and ecological analysis, there is a notable lack of dedicated, open-source tools within the Python ecosystem specifically designed for calculating these forest structural metrics from lidar data. Calculating these metrics is non-trivial, and several steps are often required to process the point clouds in order to generate these metrics. Existing open source solutions are primarily in the R programming language [@rousselLidRPackageAnalysis:2020a;  @rousselAirborneLiDARData:2024; @dealmeidaLeafRCalculatesLeaf:2021] or are proprietary, computationally intensive, or not flexible enough for the variety of ecological contexts in which these metrics are needed [@lastools; @fusion; @globalmapper]. This gap makes it difficult for researchers, ecologists, and forest managers to integrate lidar-based analysis into their workflows efficiently.
 
[…]
