+++ paper/paper.bib
@@ -1,11 +1,11 @@
 
 @article{kamoskeLeafAreaDensity2019,
-	title = {Leaf area density from airborne {LiDAR}: Comparing sensors and resolutions in a temperate broadleaf forest ecosystem},
+	title = {Leaf Area Density from Airborne {LiDAR}: Comparing Sensors and Resolutions in a Temperate Broadleaf Forest Ecosystem},
 	volume = {433},
 	issn = {03781127},
 	url = {https://linkinghub.elsevier.com/retrieve/pii/S0378112718315561},
 	doi = {10.1016/j.foreco.2018.11.017},
-	shorttitle = {Leaf area density from airborne {LiDAR}},
+	shorttitle = {Leaf Area Density From Airborne {LiDAR}},
 	abstract = {Forest processes that play an essential role in carbon sequestration, such as light use eﬃciency, photosynthetic capacity, and trace gas exchange, are closely tied to the three-dimensional structure of forest canopies. However, the vertical distribution of leaf traits is not uniform; leaves at varying vertical positions within the canopy are physiologically unique due to diﬀering light and environmental conditions, which leads to higher carbon storage than if light conditions were constant throughout the canopy. Due to this within-canopy variation, three-dimensional structural traits are critical to improving our estimates of global carbon cycling and storage by Earth system models and to better understanding the eﬀects of disturbances on carbon sequestration in forested ecosystems. In this study, we describe a reproducible and open-source methodology using the R programming language for estimating leaf area density ({LAD}; the total leaf area per unit of volume) from airborne {LiDAR}. Using this approach, we compare {LAD} estimates at the Smithsonian Environmental Research Center in Maryland, {USA}, from two airborne {LiDAR} systems, {NEON} {AOP} and {NASA} G-{LiHT}, which diﬀer in survey and instrument speciﬁcations, collections goals, and laser pulse densities. Furthermore, we address the impacts of the spatial scale of analysis as well as diﬀerences in canopy penetration and pulse density on {LAD} and leaf area index ({LAI}) estimates, while oﬀering potential solutions to enhance the accuracy of these estimates. {LAD} estimates from airborne {LiDAR} can be used to describe the three-dimensional structure of forests across entire landscapes. This information can help inform forest management and conservation decisions related to the estimation of aboveground biomass and productivity, the response of forests to large-scale disturbances, the impacts of drought on forest health, the conservation of bird habitat, as well as a host of other important forest processes and responses.},
 	pages = {364--375},
 	journaltitle = {Forest Ecology and Management},
@@ -18,13 +18,13 @@
 }
 
 @article{mcelhinnyForestWoodlandStand2005,
-	title = {Forest and woodland stand structural complexity: Its definition and measurement},
+	title = {Forest and Woodland Stand Structural Complexity: Its Definition and Measurement},
 	volume = {218},
 	rights = {https://www.elsevier.com/tdm/userlicense/1.0/},
 	issn = {03781127},
 	url = {https://linkinghub.elsevier.com/retrieve/pii/S0378112705005001},
 	doi = {10.1016/j.foreco.2005.08.034},
-	shorttitle = {Forest and woodland stand structural complexity},
+	shorttitle = {Forest and Woodland Stand Structural Complexity},
[…]
