### 2024-09-29 fe8eb53
Update references and acknowledgements in paper
Added new references and acknowledged the National Science Foundation award. This enhances the credibility of the research and indicates additional funding sources and recognition.
