+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @article{kamoskeLeafAreaDensity2019,
-	title = {Leaf Area Density from Airborne {LiDAR}: Comparing Sensors and Resolutions in a Temperate Broadleaf Forest Ecosystem},
+	title = {Leaf area density from airborne {LiDAR}: Comparing sensors and resolutions in a temperate broadleaf forest ecosystem},
 	volume = {433},
 	issn = {03781127},
 	url = {https://linkinghub.elsevier.com/retrieve/pii/S0378112718315561},
@@ -18,7 +18,7 @@
 }
 
 @article{mcelhinnyForestWoodlandStand2005,
-	title = {Forest and Woodland Stand Structural Complexity: Its Definition and Measurement},
+	title = {Forest and woodland stand structural complexity: Its definition and measurement},
 	volume = {218},
 	rights = {https://www.elsevier.com/tdm/userlicense/1.0/},
 	issn = {03781127},
@@ -38,7 +38,7 @@
 }
 
 @article{drakeEstimationTropicalForest2002,
-	title = {Estimation of Tropical Forest Structural Characteristics Using Large-footprint Lidar},
+	title = {Estimation of tropical forest structural characteristics using large-footprint lidar},
 	volume = {79},
 	rights = {https://www.elsevier.com/tdm/userlicense/1.0/},
 	issn = {00344257},
@@ -57,7 +57,7 @@
 }
 
 @report{dealmeidaLeafRCalculatesLeaf2021,
-	title = {{leafR}: Calculates the Leaf Area Index ({LAD}) and Other Related Functions},
+	title = {{leafR: Calculates the Leaf Area Index (LAD) and Other Related Functions}},
 	url = {https://CRAN.R-project.org/package=leafR},
 	type = {manual},
 	author = {de Almeida, Danilo Roberti Alves and Stark, Scott Christopher and Silva, Carlos Alberto and Hamamura, Caio and Valbuena, Ruben},
@@ -73,7 +73,7 @@
 }
 
 @article{rousselLidRPackageAnalysis2020a,
-	title = {{lidR}: An {R} Package For Analysis of Airborne Laser Scanning ({ALS}) Data},
+	title = {{lidR}: An {R} package for analysis of {Airborne Laser Scanning} ({ALS}) data},
 	volume = {251},
 	issn = {00344257},
 	url = {https://linkinghub.elsevier.com/retrieve/pii/S0034425720304314},
@@ -91,7 +91,7 @@
 }
 
 @article{hurlbertNonconceptSpeciesDiversity1971,
-	title = {The Nonconcept of Species Diversity: A Critique and Alternative Parameters},
+	title = {{The Nonconcept of Species Diversity: A Critique and Alternative Parameters}},
 	volume = {52},
 	rights = {http://onlinelibrary.wiley.com/{termsAndConditions}\#vor},
 	issn = {0012-9658, 1939-9170},
@@ -156,4 +156,100 @@
 	url = {https://research.fs.usda.gov/pnw/products/dataandtools/tools/fusion/ldv-lidar-processing-and-visualization-software-version-4.40},
 	author = {{McGaughey}, {Robert}},
 	date = {2022},
+}
+
+@article{pascualRoleImprovedGround2020,
+	title = {{The Role of Improved Ground Positioning and Forest Structural Complexity When Performing Forest Inventory Using Airborne Laser Scanning}},
+	volume = {12},
+	rights = {https://creativecommons.org/licenses/by/4.0/},
+	issn = {2072-4292},
+	url = {https://www.mdpi.com/2072-4292/12/3/413},
+	doi = {10.3390/rs12030413},
[…]
