### 2024-09-30 9147c73
Add visualizations to documentation
Inserted figures for Height Above Ground, Canopy Height Model, Plant Area Index, and voxel-based forest structure metrics into the documentation. This enhances the clarity and visual appeal of the content for users.
