+++ docs/source/metric_calculation.rst
@@ -3,6 +3,13 @@ Forest Structure Metrics Calculations
 
 The core forest structural metrics generated by PyForestScan are based on well-established methods in ecology.
 
+.. figure:: images/voxel_metrics.jpg
+   :alt: Forest structure from voxels.
+   :align: center
+   :figclass: align-center
+
+   Forest structural metrics as calculated from points within voxels. Voxel resolution is given by Δx, Δy, Δz.
+
 Canopy Height
 -------------
 
