### 2024-09-30 #7283 @iosefa
@kthyng

Ok. We have fixed the parenthetical and inline references and also checked capitalization on all references in the text. That should be fixed now. We have also added a few figures to the readthedocs for illustrative purposes. There is also a simple example in a jupyter notebook that serves as a tutorial here: https://github.com/iosefa/PyForestScan/blob/main/notebooks/demo.ipynb
Thank you!
