### 2024-10-03 #7283 @npucino
Absolutely, I will be happy to review it, thanks for the consideration .
Nick

Inviato da Outlook per Android<https://aka.ms/AAb9ysg>
