### 2024-10-04 #7314 @editorialbot
[REVIEW]: PyForestScan: A Python Library for Calculating Forest Structural Metrics from LiDAR Point Cloud Data
**Submitting author:** @iosefa (<a href="http://orcid.org/0000-0001-5941-4601">Joseph Emile Honour Percival</a>)
**Repository:** https://github.com/iosefa/PyForestScan
**Branch with paper.md** (empty if default branch): docs/joss-paper
**Version:** v0.2.2
