### 2024-10-24 #7314 @npucino
Hi, sorry the delayed review. I plan to complete it next week. Apologies for my sluggishness, regards, Nick

Nicolas Pucino (He/Him)

Postdoctoral Researcher @ Australian National University | Fenner School of Environment and Society

Visiting Scientist @ CSIRO Environment | Living Landscapes Program, Earth Observation

ISPRS SC Vice President 2022-2026 | International Society of Photogrammetry and Remote Sensing

[cid:44f3120a-f883-447b-b5c3-02e7c49b7d97]     [cid:a51fb505-95b8-494b-91ed-d99ea9dc956b]      [cid:12a0d605-e600-44d3-94b7-771611214491]
