### 2024-10-24 iosefa/PyForestScan#11 @giswqs
JOSS Paper Suggestions
Re: https://github.com/openjournals/joss-reviews/issues/7314

I have a few suggestions regarding the JOSS paper:

1. **Title Consistency:** The title uses "LiDAR," while the main text refers to it as "lidar." Please ensure consistency throughout.
2. **Term Capitalization:** In the summary, terms like "foliage height diversity" (FHD) are presented in all lowercase, whereas in the Statement of Need, they are capitalized as "Foliage Height Diversity" (FHD). The same inconsistency applies to PAD and PAI. Please standardize the capitalization across the document.
3. **Citation Recommendation:** Consider citing PDAL: https://doi.org/10.5281/zenodo.2616780.
