### 2024-10-25 iosefa/PyForestScan#12 @giswqs
JOSS Review Feedback
Re: https://github.com/openjournals/joss-reviews/issues/7314

## Missing Items

Two critical elements are currently absent:

- **Automated Tests**: There are no automated tests to verify the software's functionality and Python version support (3.10-3.13)
- **Community Guidelines**: While a CONTRIBUTING.md file exists in the repository, no  community guidelines are available on the documentation website.

**JOSS Reivew Checklists**

## Areas for Improvement

To enhance user experience, consider the following recommendations:

1. **Sample Datasets and Jupyter Notebooks**: Providing sample datasets along with complete Jupyter Notebook examples would facilitate testing of the package's functionality.
2. **Binder Integration**: Link to [mybinder.org](https://mybinder.org) to allow users to try out notebooks without local installation. Although installing PDAL on Google Colab can be challenging, using a Dockerfile to set up the environment for Binder could streamline this process.
3. **Data Visualization Examples**: The package includes data visualization features that are not showcased in the Quick Start Example.
4. **Reference to PDAL**: Enhance the references section by citing PDAL (https://doi.org/10.5281/zenodo.2616780), as it is a crucial dependency.

**JOSS Reivew Checklists**

## Request for Clarification

It appears that the submitting author is the sole contributor of source code to this repository. Clarification on the roles of the two co-authors in the development of this package would be beneficial.

**JOSS Reivew Checklists**

- [ ] **Contribution and Authorship**: Has the submitting author (@iosefa) made major contributions to the software? Does the list of paper authors seem appropriate and complete?

## Repository Suggestions

1. **Enhance Repository Information**: Add a project description, website, and relevant tags to the About section (next to the green Code button).
2. **Include Badges**: Add badges (e.g., PyPI, Binder) to the README.md for better visibility and usability.

---
You can refer to the[ segment-geospatial](https://github.com/opengeos/segment-geospatial) Python package for implementing these suggested changes.
