### 2024-11-05 iosefa/PyForestScan#13 @npucino
JOSS Paper review by Nick
Here are my main suggestions to improve the **PyForestScan software paper** for [JOSS submission](https://github.com/openjournals/joss-reviews/issues/7314).

### Summary Section
- **Clarify Acronym**: Before mentioning LiDAR as an acronym, spell it out in full as *LIght Detection And Ranging*, similar to how you've done for *FHD*, *PAD*, and *PAI*.
- **Digital Terrain Models**: Add the term “models” after "digital terrain."

### State of the field
- **Missing**: This section is completely missing from the paper. Here a list of previous work or alternative software should be presented in a way to then reinforce the need for PyForestScan.

### Statement of Need
- **Data Complexity and Size**:
  - **Complexity**: Please specify the complexities that current methods struggle to address and how *PyForestScan* aims to solve these issues.
  - **Size**: Could you provide more specific details when discussing size? Typically, LiDAR data is processed in tiles. For example, 1km x 1km LiDAR tiles with low average point densities (2-3 pts/m²) are usually around 20-30 MB, which is manageable. However, with higher densities (e.g., UAV LiDAR data at 2000-3000 pts/m²), file sizes can become a limiting factor. Please clarify the specific size challenges.

- **Last Sentence of First Section - LiDAR-based Analysis**:
  - While “LiDAR-based analysis” is generally understood, it's more precise here to refer to "point-cloud-based analysis" rather than LiDAR-based analysis. LiDAR is the sensor technique that produces point clouds, often discretized into rasters. Since this section discusses LiDAR-derived point-cloud analysis (as mentioned earlier), consider using "point-cloud-based analysis."

- **Second Section - LiDAR Data**:
  - See the previous point. It seems you are specifically referring to point-cloud data here.

- **Points Generated from SfM**:
  - How are PAD, PAI, and FHD metrics derived from airborne SfM data, given that aerial SfM typically captures only the canopy top without penetrating vegetation layers? Without integrating terrestrial images or using off-nadir, low-altitude images, SfM can only capture the top layer, making voxel creation impossible. Of course, SfM CAN generate 3D point clouds of trees, but large-scale forest point clouds (with tiepoints derived from several heights along the vegetation profile) from aerial SfM seem not feasible or feasible only with expensive drone swarms flying through the canopy. Please clarify how SfM data can be used in your package to derive these metrics.

### Acknowledgment
- **Test dataset to validate and benchmark PyForestScan**:
  - It is mentioned that a single noise-free classified point cloud has been used for testing and validating the package. It would be valuable to know more about the data used for testing in this paper, without having to find the information in the papers cited in the test_data folder. Please add some information in this paper, such as (1) location, (2) dominant ecosystem/vegetation type/vegetation structure, (3) point cloud density and (4) preprocessing steps if any.
