### 2024-11-05 iosefa/PyForestScan#14 @npucino
JOSS Review Feedback by Nick
Here is my feedback to improve the PyForestScan software for https://github.com/openjournals/joss-reviews/issues/7314.

### General checks

- [x] **Contribution and authorship: Has the submitting author ([@iosefa](https://github.com/iosefa)) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?**
The main author is the only contributor in the source code of this package based on the commits history. However, the co-authors might have helped the development in early stages before using git to track changes. As it stands, without a few lines of explanation of their roles, I assume the other co-authors had mentoring, reviewing and funding acquisition roles. In JOSS, **these roles do not qualify as authorship** ([link](https://joss.readthedocs.io/en/latest/submitting.html#authorship)) and should not be included in this submission. Please clarify the co-authors role or do not include them.

- [x] **Substantial scholarly effort: Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)**
While this package does simplify several tasks (reading, cleaning, classifying, filtering, digital terrain and canopy height model creations and more) by acting as a wrapper around PDAL—which can be challenging to master due to its syntax and pipeline structure—it does not introduce any substantial new functionality except the computation of the metrics. For example, in the benchmarking section of the documentation, claims such as, “PyForestScan is designed for high performance and memory efficiency, ensuring it can handle large-scale point cloud datasets effectively,” could equally apply to PDAL itself, as PyForestScan primarily repackages PDAL’s capabilities rather than adding novel performance enhancements.
In my opinion, this package is indeed valuable for those looking to streamline tasks that might otherwise require working with PDAL’s more complex syntax with the added bonus of including few key tree metrics. However, since it functions as a wrapper rather than a significant software innovation, I believe it does not meet the minimum requirements for publication as original software in JOSS.

### Functionality

- [x] **Installation:**
Installation works well, however, as you provide a Jupyter Notebook as a demo, you could include in the quickstart guide a section on how to install jupyter notebook (`conda install jupyter`) and run the demo with your test data.

- [x] **Functionality:**
In the demo notebook, there is an error importing plot_3d simply due to this functionality being removed due to dependencies problem ([commit 249ca40](https://github.com/iosefa/PyForestScan/commit/249ca4058d444f140ccf0fdd04fcd6ba1a14a639)). You just forgot to remove it from the demo notebook.
_[OPTIONAL]_ However, I think it is worthwhile to have a 3D visualization capability, so if you can fix the dependencies problem you had, it would be cool to re-integrate this function and provide an extra cell in the demo notebook where you showcase plot_3d function and output. **Just a suggestion, not required.**

- [x] **Performance:**
The benchmark provided in the documentation appears to compare PDAL with `leafR `rather than showcasing any substantial improvements introduced by this package itself.  This because the bulk of the processing time is spent in reading the point clouds rather than computing the metrics, I believe. As a consequence, there doesn’t seem to be a specific contribution from the authors that significantly speeds up the PAI computation beyond what PDAL (repackaged into PyForestScan) already offers.

### Documentation
- [x] **Example usage:** As you have prepared a demo notebook, please mention the demo notebook from the documentation.

### Software paper

- [x] **References:**
Citation of PDAL is missing and of uttermost importance given this package almost entirely relies on PDAL.

For more detailed paper feedback, refer [to this issue I opened.](https://github.com/iosefa/PyForestScan/issues/13)
