### 2024-11-05 #7314 @npucino
@mengqi-z I apologize for the time I took.
Here is my feedback:

- [Paper comments](https://github.com/iosefa/PyForestScan/issues/13)
- [Overall submission comments](https://github.com/iosefa/PyForestScan/issues/14)

My main concern is on the _substantial scholarly effort_ point, all the rest can be easily improved.
