### 2024-11-05 #7314 @mengqi-z
@npucino - Thank you for your time and effort in reviewing the paper!

@iosefa - Along with addressing the comments, could you please provide a detailed justification in response to @npucino's concern regarding the substantial scholarly effort? You may refer to the JOSS guidelines on [Substantial Scholarly Effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) for additional context.
