### 2024-11-14 #7314 @iosefa
@giswqs @npucino @mengqi-z
Thank you very much for your detailed reviews and feedback. I decided to wait until hearing the comments from @npucino before making any changes to the manuscript and codebase. Now that I’ve received comments from @npucino, I will move forward and prepare a detailed response. I’m currently a bit tied up, but I wanted to acknowledge your efforts and let you know that I’ll respond as soon as possible. Thank you again for your time and thoughtful input. I’ll be in touch soon.
