### 2024-11-21 8d1f7fc
Correct typos and standardize term capitalization for #11
Update the capitalization of "Python" and "lidar" in the title and main text for consistency. Ensure all metric names like Plant Area Index (PAI) and Foliage Height Diversity (FHD) follow proper capitalization to enhance readability and professionalism.
