### 2024-11-21 24b6a23
Add PDAL citation to paper references for #11.
This update includes the citation for the PDAL library in the references section of the paper. The citation is added for the mention of PDAL in the summary and statement of need sections, providing proper attribution to the contributing authors and source.
