+++ paper/paper.md
@@ -32,13 +32,13 @@ bibliography: paper.bib
 
 # Summary
 
-PyForestScan is an open-source Python library designed for calculating forest structural metrics from lidar point cloud data at scale. The software supports input formats including .las, .laz, and .copc files, efficiently handles large-scale lidar datasets, and calculates key ecological metrics such as foliage height diversity (FHD), plant area density (PAD), canopy height, and plant area index (PAI). In addition to metrics computation, the library supports the import and generation of digital terrain models, the generation of GeoTIFF outputs, and integrates with geospatial libraries like PDAL, making it a valuable tool for forest monitoring, carbon accounting, and ecological research.
+PyForestScan is an open-source Python library designed for calculating forest structural metrics from lidar point cloud data at scale. The software supports input formats including .las, .laz, and .copc files, efficiently handles large-scale lidar datasets, and calculates key ecological metrics such as foliage height diversity (FHD), plant area density (PAD), canopy height, and plant area index (PAI). In addition to metrics computation, the library supports the import and generation of digital terrain models, the generation of GeoTIFF outputs, and integrates with geospatial libraries like PDAL [@howard_butler_2024_13993879], making it a valuable tool for forest monitoring, carbon accounting, and ecological research.
 
 # Statement of Need
 
 Remote sensing data, particularly Light Detection and Ranging (lidar) data from airborne sensors, are becoming increasingly accessible, offering a detailed understanding of forest ecosystems at fine spatial resolutions. This data is critical for calculating forest structural metrics such as canopy height, canopy cover, plant area index (PAI), plant area density (PAD), foliage height diversity (FHD), as well as digital terrain models (DTM), which are essential for forest management, biodiversity conservation, and carbon accounting [@mcelhinnyForestWoodlandStand2005; @drakeEstimationTropicalForest2002; @pascualRoleImprovedGround2020; @guerra-hernandezUsingBitemporalALS2024; @pascualIntegratedAssessmentCarbon2023; @pascualNewRemoteSensingbased2021a]. However, working with large-scale lidar datasets remains a challenge due to the complexity and size of the data. Additionally, despite Python being a powerful language widely used for geospatial and ecological analysis, there is a notable lack of dedicated, open-source tools within the Python ecosystem specifically designed for calculating these forest structural metrics from lidar data. Calculating these metrics is non-trivial, and several steps are often required to process the point clouds in order to generate these metrics. Existing open source solutions are primarily in the R programming language [@rousselLidRPackageAnalysis2020a;  @rousselAirborneLiDARData2024; @dealmeidaLeafRCalculatesLeaf2021] or are proprietary, computationally intensive, or not flexible enough for the variety of ecological contexts in which these metrics are needed [@lastools; @fusion; @globalmapper]. This gap makes it difficult for researchers, ecologists, and forest managers to integrate lidar-based analysis into their workflows efficiently.
 
[…]
