+++ paper/paper.bib
@@ -253,3 +253,42 @@ Use of these data requires citation of this dataset},
 	urldate = {2024-09-29},
 	date = {2024-08-01},
 }
+
+@software{howard_butler_2024_13993879,
+  author       = {Howard Butler and
+                  Andrew Bell and
+                  Michael P. Gerlek and
+                  chambbj and
+                  Pete Gadomski and
+                  Connor Manning and
+                  Mateusz Łoskot and
+                  Bas Couwenberg and
+                  Norman Barker and
+                  Paul Ramsey and
+                  Julia Dark and
+                  Kyle Mann and
+                  Nicolas Chaulet and
+                  Even Rouault and
+                  Guilhem Villemin and
+                  Claire Foster and
+                  Ogi Moore and
+                  Michael Rosen and
+                  Scott Lewis and
+                  Grigory and
+                  Kirk McKelvey and
+                  Martin Dobias and
+                  Michael D. Smith and
+                  Bram and
+                  xantares and
+                  Vicky Vergara and
+                  Robert Coup and
+                  Kristian Evers and
+                  Daniel Brookes},
+  title        = {PDAL/PDAL: 2.8.1},
+  month        = oct,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {2.8.1},
+  doi          = {10.5281/zenodo.13993879},
+  url          = {https://doi.org/10.5281/zenodo.13993879}
+}
