### 2024-11-27 16bcad1
Remove author and update paper content to address reviewer comments.
Removed Ryan Perroy from author list. Enhanced the summary and statement of need sections to clarify functionalities and emphasize forest structure metric calculation. Clarified the challenges associated with size and the features that the library provides. Added new sections: usage and contributing.
