+++ paper/paper.bib
@@ -292,3 +292,77 @@ Use of these data requires citation of this dataset},
   doi          = {10.5281/zenodo.13993879},
   url          = {https://doi.org/10.5281/zenodo.13993879}
 }
+
+@software{manning_entwine,
+  author       = {Connor Manning},
+  title        = {Entwine: Open Source Point Cloud Indexing},
+  year         = {2024},
+  url          = {https://entwine.io/},
+  note         = {Latest version accessed on 2024-11-22}
+}
+
+@article{doi:10.1111/2041-210X.13901,
+author = {Borowiec, Marek L. and Dikow, Rebecca B. and Frandsen, Paul B. and McKeeken, Alexander and Valentini, Gabriele and White, Alexander E.},
+title = {Deep learning as a tool for ecology and evolution},
+journal = {Methods in Ecology and Evolution},
+volume = {13},
+number = {8},
+pages = {1640-1660},
+keywords = {artificial intelligence, automation, computer vision, machine learning, modelling, neural networks, statistics},
+doi = {https://doi.org/10.1111/2041-210X.13901},
+url = {https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/2041-210X.13901},
+eprint = {https://besjournals.onlinelibrary.wiley.com/doi/pdf/10.1111/2041-210X.13901},
+abstract = {Abstract Deep learning is driving recent advances behind many everyday technologies, including speech and image recognition, natural language processing and autonomous driving. It is also gaining popularity in biology, where it has been used for automated species identification, environmental monitoring, ecological modelling, behavioural studies, DNA sequencing and population genetics and phylogenetics, among other applications. Deep learning relies on artificial neural networks for predictive modelling and excels at recognizing complex patterns. In this review we synthesize 818 studies using deep learning in the context of ecology and evolution to give a discipline-wide perspective necessary to promote a rethinking of inference approaches in the field. We provide an introduction to machine learning and contrast it with mechanistic inference, followed by a gentle primer on deep learning. We review the applications of deep learning in ecology and evolution and discuss its limitations and efforts to overcome them. We also provide a practical primer for biologists interested in including deep learning in their toolkit and identify its possible future applications. We find that deep learning is being rapidly adopted in ecology and evolution, with 589 studies (64\%) published since the beginning of 2019. Most use convolutional neural networks (496 studies) and supervised learning for image identification but also for tasks using molecular data, sounds, environmental data or video as input. More sophisticated uses of deep learning in biology are also beginning to appear. Operating within the machine learning paradigm, deep learning can be viewed as an alternative to mechanistic modelling. It has desirable properties of good performance and scaling with increasing complexity, while posing unique challenges such as sensitivity to bias in input data. We expect that rapid adoption of deep learning in ecology and evolution will continue, especially in automation of biodiversity monitoring and discovery and inference from genetic data. Increased use of unsupervised learning for discovery and visualization of clusters and gaps, simplification of multi-step analysis pipelines, and integration of machine learning into graduate and postgraduate training are all likely in the near future.},
+year = {2022}
+}
+
+@article{doi:10.1111/2041-210X.14040,
+author = {Atkins, Jeff W. and Costanza, Jennifer and Dahlin, Kyla M. and Dannenberg, Matthew P. and Elmore, Andrew J. and Fitzpatrick, Matthew C. and Hakkenberg, Christopher R. and Hardiman, Brady S. and Kamoske, Aaron and LaRue, Elizabeth A. and Silva, Carlos Alberto and Stovall, Atticus E. L. and Tielens, Elske K.},
+title = {Scale dependency of lidar-derived forest structural diversity},
+journal = {Methods in Ecology and Evolution},
+volume = {14},
+number = {2},
+pages = {708-723},
[…]
