### 2024-11-30 04cd7b3
Refactor documentation structure and remove legacy files
Reorganize image files into a more logical directory hierarchy under `docs/images` for improved maintainability. Remove the obsolete ReadTheDocs configuration and unnecessary documentation files, such as `api_reference.rst`, `benchmarks.rst`, `conf.py`, and `demo.ipynb`, to streamline the project. The changes aim to enhance the clarity and efficiency of the documentation build process.
