+++ docs/requirements.txt
@@ -1,11 +1,0 @@
-sphinx>=7.4.0
-sphinx_rtd_theme>=1.3.0
-sphinx-autodoc-typehints>=1.19.6
-numpy>=1.25.1
-pandas>=2.0.3
-rasterio>=1.3.8
-geopandas>=0.13.2
-shapely>=2.0.1
-pyproj>=3.6.0
-matplotlib>=3.7.2
-scipy>=1.10.1
