+++ docs/api.md
@@ -1,0 +1,3 @@
+# API Documentation
+
+::: pyforestscan.calculate
