+++ docs/source/index.rst
@@ -1,28 +1,0 @@
-.. pyforestscan documentation master file, created by
-   sphinx-quickstart on Thu Sep 14 16:06:59 2023.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-pyforestscan
-============
-
-Welcome to pyforestscan's documentation!
-
-Contents:
----------
-
-.. toctree::
-   :maxdepth: 3
-
-   introduction
-   getting_started
-   metric_calculation
-   benchmarks
-   api_reference
-
-Indices and Tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`

[…]
