+++ Dockerfile
@@ -2,14 +2,20 @@ FROM jupyter/base-notebook:latest
 LABEL maintainer="Iosefa Percival"
 LABEL repo="https://github.com/iosefa/PyForestScan"
 
-USER 1000
+USER root
+RUN apt-get update -y && apt-get install -y \
+    gcc g++ make \
+    libgdal-dev libgl1 sqlite3 && \
+    apt-get clean && rm -rf /var/lib/apt/lists/*
 
-RUN mamba install -c conda-forge pdal gdal -y && \
-    pip install -U pyforestscan jupyter-server-proxy && \
-    mamba update -c conda-forge -y && \
-    jupyter server extension enable --sys-prefix jupyter_server_proxy && \
-    fix-permissions "${CONDA_DIR}" && \
+RUN fix-permissions "${CONDA_DIR}" && \
     fix-permissions "/home/${NB_USER}"
+
+USER 1000
+RUN mamba install -c conda-forge sqlite gdal pdal -y && \
+    pip install --no-cache-dir pyforestscan jupyter-server-proxy && \
+    mamba update -c conda-forge -y && \
+    jupyter server extension enable --sys-prefix jupyter_server_proxy
 
 RUN mkdir ./examples
 COPY /docs/examples ./examples
