### 2024-11-30 63fe3e6
Refactor contributing docs and update dependencies
Moved CONTRIBUTING.md to docs/contributing.md for better organization and added a new example notebook for large point clouds processing. Updated the README to reflect the added GDAL dependency. Enhanced index.md with clearer package features and simplified forest metrics processing example.
