### 2024-11-30 82a9d0c
update function for calculating metrics and plotting PAI
