### 2024-11-30 cc7637f
Update jupyter notebooks, as well as generate_dtm().
