### 2024-12-01 a565239
Add optional metric_name for plot colorbar labels
The `plot_metric` function now includes an optional `metric_name` parameter to specify the label for the colorbar. If `metric_name` is not provided, the plot title will be used as the default label. This enhancement allows for more descriptive or unit-specific labels on the colorbar.
