### 2024-12-01 7ae8e95
Refactor documentation structure and content
The documentation for pyforestscan has been restructured for better organization and clarity. New sections are added for specific topics like Canopy Height Models, Digital Terrain Models, etc. This refactor enhances accessibility and usability for users seeking particular information.
