+++ docs/usage/forest-structure/chm.md
@@ -1,0 +1,17 @@
+# Canopy Height Models
+
+Canopy height is given as the maximum height above ground for a point within each grid cell or for all points within a given polygon.
+
+$$H_{\text{canopy}} = \max(HAG_{\text{points}})$$
+
+Where:
+
+-   $H_{\text{canopy}}$ is the canopy height for the grid cell or polygon.
+-   $HAG_{\text{points}}$ represents the set of heights above ground of all points within a given grid cell or polygon.
+
+To calculate canopy height and generate a canopy height model, we first import the `caulculate_chm` function
+
+## Gridded CHM
+
+## Canopy Height for Abstract Polygons
+
