+++ README.md
@@ -5,20 +5,19 @@
 [![Tests](https://img.shields.io/github/actions/workflow/status/iosefa/PyForestScan/main.yml?branch=main)](https://github.com/iosefa/PyForestScan/actions/workflows/main.yml)
 [![Coverage](https://img.shields.io/codecov/c/github/iosefa/PyForestScan/main)](https://codecov.io/gh/iosefa/PyForestScan)
 
+**Calculate Forest Structural Metrics from lidar point clouds in Python**
+
 ![Height Above Ground](./screenshots/hag.png)
 
 ## Overview
 
-PyForestScan is a Python library designed for analyzing and visualizing forest structure using airborne 
-3D point cloud data. The library helps derive important forest metrics such as Canopy Height, 
-Plant Area Index (PAI), Canopy Cover, Plant Area Density (PAD), and Foliage Height Diversity (FHD).
+PyForestScan is a Python library designed for analyzing and visualizing forest structure using airborne 3D point cloud data. The library helps derive important forest metrics such as Canopy Height, Plant Area Index (PAI), Canopy Cover, Plant Area Density (PAD), and Foliage Height Diversity (FHD).
 
 ## Features
 
 - **Forest Metrics**: Calculate and visualize key metrics like Canopy Height, PAI, PAD, and FHD.
 - **Large Point Cloud Support**: Utilizes efficient data formats such as EPT for large point cloud processing.
-- **Airborne Data Compatibility**: Supports LiDAR and Structure from Motion (SfM) data from drones and UAVs.
-- **Visualization**: Create 2D and 3D visualizations of forest structures.
+- **Visualization**: Create 2D and 3D visualizations of forest structure and structural metrics
 - **Extensibility**: Easily add custom filters and visualization techniques to suit your needs.
 
 ## Installation
