+++ docs/usage/forest-structure/intro.md
@@ -1,0 +1,18 @@
+# Introduction
+
+The core forest structural metrics generated by PyForestScan are based
+on well-established methods in ecology.
+
+![Forest structural metrics as calculated from points within voxels. Voxel resolution is given by Δx, Δy, Δz.](../../images/voxel_metrics.jpg)
+<figcaption align="center">Forest structural metrics as calculated from points within voxels. Voxel resolution is given by Δx, Δy, Δz.</figcaption>
+
+The calculation follows the method outlined in Kamoske et al. 2019.
+
+
+## References
+
+Kamoske, Aaron G., Kyla M. Dahlin, Scott C. Stark, and Shawn P. Serbin.
+2019. "Leaf Area Density from Airborne LiDAR: Comparing Sensors and
+Resolutions in a Temperate Broadleaf Forest Ecosystem." Forest Ecology
+and Management 433 (February): 364--75.
+<https://doi.org/10.1016/j.foreco.2018.11.017>.

[…]
