+++ requirements.txt
@@ -8,3 +8,4 @@ pandas~=2.2.2
 numpy~=2.1.1
 matplotlib~=3.9.2
 scipy~=1.14.1
+tqdm~=4.66.5
