### 2024-12-01 dd8e9b9
Update contributions and usage section
