### 2024-12-03 799719e
Add new images and documentation files
This commit introduces several new files, including images, example data, and API documentation for various modules. Additionally, a Code of Conduct is established, fostering a welcoming community environment. The example notebook on calculating forest metrics has been updated to improve clarity and efficiency in demonstrating the use of the PyForestScan library.
