+++ docs/usage/forest-structure/intro.md
@@ -1,13 +1,18 @@
 # Introduction
 
-The core forest structural metrics generated by PyForestScan are based
-on well-established methods in ecology.
+The core forest structural metrics generated by PyForestScan are based on well-established methods in ecology.
 
 ![Forest structural metrics as calculated from points within voxels. Voxel resolution is given by Δx, Δy, Δz.](../../images/voxel_metrics.jpg)
 <figcaption align="center">Forest structural metrics as calculated from points within voxels. Voxel resolution is given by Δx, Δy, Δz.</figcaption>
 
-The calculation follows the method outlined in Kamoske et al. 2019.
+Most calculations for the metrics follows the method outlined in Kamoske et al. 2019.
 
+## Metrics
+
+* [Canopy Height Models (CHM)](chm.md)
+* [Plant Area Density (PAD)](pad.md)
+* [Plant Area Index (PAI)](pai.md)
+* [Foliage Height Diversity (FHD)](fhd.md)
 
 ## References
 
