+++ docs/benchmarks.md
@@ -1,24 +1,13 @@
 # Benchmarking
 
-PyForestScan is designed for high performance and memory efficiency,
-ensuring it can handle large-scale point cloud datasets effectively.
-While no other Python libraries specifically calculate these forest
-structure metrics, there are alternatives in R, such as the
-`leafR` library (Almeida et al. 2021), that offer similar
-functionality.
+PyForestScan is designed for high performance and memory efficiency, ensuring it can handle large-scale point cloud datasets effectively. While no other Python libraries specifically calculate these forest structure metrics, there are alternatives in R, such as the`leafR` library (Almeida et al. 2021), that offer similar functionality.
 
-We provide a direct performance comparison between PyForestScan and
-`leafR` to demonstrate its efficiency. In both cases, we
-calculate Plant Area Index (PAI) (this is labelled as Leaf Area Index in
-the `leafR` library) on a LAS tile, repeating the process
-100 times and plotting the results. The measured ...
+We provide a direct performance comparison between PyForestScan and`leafR` to demonstrate its efficiency. In both cases, we calculate Plant Area Index (PAI) (this is labelled as Leaf Area Index in the `leafR` library) on a LAS tile, repeating the process 100 times and plotting the results. The measured benchmarking is only done on the functions to calculate PAI/LAI. It does not include time taken to load the point cloud, etc.
 
 
-The benchmarks were conducted on a Mac with an Apple M3 Max processor
-(16 cores) and 128GB RAM.
+The benchmarks were conducted on a Mac with an Apple M3 Max processor (16 cores) and 128GB RAM.
 
-![Benchmark comparison between PyForestScan and leafR](images/lai_computation_times.png){.align-center
-width="600px"}
+![Benchmark comparison between PyForestScan and leafR](images/lai_computation_times.png)
 
 ## Code Used
 To calculate LAI in `leafR`, and compare these with the python code, we:
@@ -79,7 +68,4 @@ unlink(temp_las_file)
 
 ## Reference
 
-Almeida, Danilo Roberti Alves de, Scott Christopher Stark, Carlos
-Alberto Silva, Caio Hamamura, and Ruben Valbuena. 2021. "leafR:
-Calculates the Leaf Area Index (LAD) and Other Related Functions."
-Manual. <https://CRAN.R-project.org/package=leafR>.
+Almeida, Danilo Roberti Alves de, Scott Christopher Stark, Carlos Alberto Silva, Caio Hamamura, and Ruben Valbuena. 2021. "leafR: Calculates the Leaf Area Index (LAD) and Other Related Functions."Manual. <https://CRAN.R-project.org/package=leafR>.
