+++ docs/usage/forest-structure/pad.md
@@ -1,82 +1,8 @@
-# Importing, Preprocessing, and Writing Data
+# Plant Area Density (PAD)
 
-To use pyforestscan, first import it in your Python project:
+## Theory
 
-```python
-import pyforestscan
-```
-
-Then, you can use it to load point cloud data and extract forest structure metrics. 
-
-The following sections will provide an overview of usage of the major functions of pyforestscan. For a complete reference of all functions in pyforestscan, please check the [API documentation](api.md). For comprehensive examples of these functions, please see the [example jupyter notebooks](examples). 
-
-## Importing Point Cloud Data
-
-pyforestscan supports reading from the following point cloud data formats:
-
-* las
-* laz
-* copc
-* ept
-
-and reading point clouds is done using the `read_lidar` function:
-
-```python
-from pyforestscan.handlers import read_lidar
-
-file_path = "../example_data/20191210_5QKB020880.laz"
-arrays = read_lidar(file_path, "EPSG:32605", hag=True)
-pointcloud = arrays[0]
-```
-
-## Preprocessing Point Cloud Data
-
-pyforestscan provides some basic functionality to help preprocess point cloud data. Many of these functions are wrapped PDAL routines. For example, to remove outliers and classify ground points:
-
-```python
-from pyforestscan.filters import remove_outliers_and_clean, classify_ground_points
-
-cleaned_arrays = remove_outliers_and_clean(arrays, mean_k=8, multiplier=3.0)
-classified_arrays = classify_ground_points(cleaned_arrays)
-```
-
-## Exporting Point Clouds
-
-pyforestscan supports exporting processed point clouds to las and laz formats. To export a point cloud as a LAZ file:
-
-```python
-from pyforestscan.handlers import write_las
-
-write_las(classified_arrays, "/path/to/exported_file.las", srs="EPSG:32605", compress=True)
-```
-
-
-# Forest Structure Metrics
-
-The core forest structural metrics generated by PyForestScan are based
-on well-established methods in ecology.
-
-![Forest structural metrics as calculated from points within voxels. Voxel resolution is given by Δx, Δy, Δz.](images/voxel_metrics.jpg)
-<figcaption align="center">Forest structural metrics as calculated from points within voxels. Voxel resolution is given by Δx, Δy, Δz.</figcaption>
-
-## Canopy Height
-
-Canopy height is given as the maximum height above ground for a point
-within each grid cell.
-
-$$H_{\text{canopy}} = \max(HAG_{\text{points}})$$
-
-Where:
-
--   `H_{\text{canopy}}` is the canopy height for the grid cell.
--   `HAG_{\text{points}}` represents the set of heights above ground of
-    all points within a given grid cell.
-
-## Plant Area Density (PAD)
-
-Plant Area Density (PAD) is a measure of the amount of plant material in
-a vertical slice of the forest, derived from airborne LiDAR data. The
-calculation follows the method outlined in Kamoske et al. 2019.
+Plant Area Density (PAD) is a measure of the amount of plant material in a vertical slice of the forest, derived from airborne LiDAR data. The calculation follows the method outlined in Kamoske et al. 2019.
 
 $$PAD_{i-1,i} = \ln\left(\frac{S_e}{S_t}\right) \frac{1}{k \Delta z}$$
 
@@ -89,60 +15,36 @@ Where:
 -   ( k ) is the extinction coefficient from the Beer-Lambert Law.
 -   ( Delta z ) is the height of each voxel.
 
-The equation calculates the natural logarithm of the ratio of entering
[…]
