### 2024-12-03 2634fdb
Add CNAME, update dependencies, and CI improvements
Introduce a CNAME for custom domain hosting and add necessary documentation dependencies. Update the GitHub Actions workflow to support multiple Python versions for testing. Increment the version from 0.1.7 to 0.2.0 to reflect these changes.
