### 2024-12-03 14cf768
Enhance CHM interpolation logic and update tests.
The interpolation parameter now supports a 'None' option, ensuring no interpolation is performed if desired. Adjusted the tests accordingly to accommodate the changes in CHM calculation, including verification for voxel processing and grid orientation. Additionally, revised test cases to handle larger datasets and improved CHM height checks.
