+++ tests/test_calculate.py
@@ -392,7 +392,6 @@ def test_calculate_fhd_partial_zero_distribution():
     fhd = calculate_fhd(voxel_returns)
     assert isinstance(fhd, np.ndarray)
     assert fhd.shape == (5, 5)
-    # Entropy should be lower where most returns are zero
     assert fhd[0, 0] < fhd[1, 1]
 
 
@@ -442,7 +441,7 @@ def test_calculate_chm_varying_heights():
     arr = create_point_cloud(num_points=100)
     arr['HeightAboveGround'] = np.random.uniform(0, 50, 100)
     voxel_resolution = (10.0, 10.0)
-    chm, extent = calculate_chm(arr, voxel_resolution)
+    chm, extent = calculate_chm(arr, voxel_resolution, interpolation=None)
 
     assert isinstance(chm, np.ndarray), "CHM should be a NumPy array."
     assert chm.ndim == 2, f"CHM should be 2-dimensional, got {chm.ndim} dimensions."
@@ -453,18 +452,19 @@ def test_calculate_chm_varying_heights():
         for j in range(chm.shape[1]):
             x_min = extent[0] + i * voxel_resolution[0]
             x_max = extent[0] + (i + 1) * voxel_resolution[0]
-            y_min = extent[2] + j * voxel_resolution[1]
-            y_max = extent[2] + (j + 1) * voxel_resolution[1]
+            y_index = chm.shape[1] - 1 - j
+            y_min = extent[2] + y_index * voxel_resolution[1]
+            y_max = extent[2] + (y_index + 1) * voxel_resolution[1]
 
             points_in_voxel = arr[
                 (arr['X'] >= x_min) & (arr['X'] < x_max) &
                 (arr['Y'] >= y_min) & (arr['Y'] < y_max)
-                ]
+            ]
 
             if len(points_in_voxel) > 0:
                 expected_max = points_in_voxel['HeightAboveGround'].max()
                 actual_value = chm[i, j]
-                assert actual_value == expected_max, (
+                assert np.isclose(actual_value, expected_max, atol=1e-6), (
                     f"CHM[{i}, {j}] should be {expected_max}, got {actual_value}."
                 )
             else:
@@ -473,48 +473,39 @@ def test_calculate_chm_varying_heights():
                     f"CHM[{i}, {j}] should be NaN or 0, got {actual_value}."
                 )
 
-
 def test_calculate_chm_negative_heights():
     """
     Test that calculate_chm correctly assigns maximum heights to each voxel,
-    including negative heights (as strange as that sounds!). For voxels with no points, CHM should be either NaN or 0.
+    including negative heights. For voxels with no points, CHM should be either NaN or 0.
     """
     np.random.seed(42)
     arr = create_point_cloud(num_points=100)
     arr['HeightAboveGround'] = np.random.uniform(-20, 50, 100)
     voxel_resolution = (10.0, 10.0)
-    chm, extent = calculate_chm(arr, voxel_resolution)
+    chm, extent = calculate_chm(arr, voxel_resolution, interpolation=None)
 
     assert isinstance(chm, np.ndarray), "CHM should be a NumPy array."
     assert chm.ndim == 2, f"CHM should be 2-dimensional, got {chm.ndim} dimensions."
     assert isinstance(extent, list), "Extent should be a list."
     assert len(extent) == 4, f"Extent should have 4 elements, got {len(extent)}."
 
-    x_min, x_max, y_min, y_max = extent
-    x_bins = np.arange(x_min, x_max + voxel_resolution[0], voxel_resolution[0])
-    y_bins = np.arange(y_min, y_max + voxel_resolution[1], voxel_resolution[1])
-
[…]
