### 2024-12-03 f8ef6bf
Bump version to 0.2.1 in setup.py.
This update reflects minor changes or improvements to the package. Ensure dependencies and configurations are compatible with this version update.
