### 2024-12-03 cf835bd
Remove unused notebook and update Binder links.
Deleted the empty 'Untitled.ipynb' file from the docs/examples directory as it contained no code or markdown content. Updated the Binder links in 'index.md' and 'README.md' to point to the correct getting-started notebook path. This helps streamline the repository and improves documentation accuracy.
