### 2024-12-03 0e70339
Add Docker image publishing to workflow
This update enhances the existing GitHub Actions workflow by introducing steps to build, tag, and push a Docker image alongside the Python package. The Docker image is built from the repository, tagged, and pushed to Docker Hub, ensuring continuous deployment and consistency across environments.
