+++ .github/workflows/publish.yml
@@ -1,4 +1,4 @@
-name: Publish Python Package
+name: Publish Python Package and Docker Image
 
 on:
   push:
@@ -29,3 +29,26 @@ jobs:
         run: |
           python setup.py sdist bdist_wheel
           twine upload dist/* -u __token__ -p $PYPI_TOKEN
+
+  docker:
+    runs-on: ubuntu-latest
+    needs: release
+
+    steps:
+      - name: Check out the repository
+        uses: actions/checkout@v3
+
+      - name: Log in to Docker Hub
+        uses: docker/login-action@v2
+        with:
+          username: ${{ secrets.DOCKER_USERNAME }}
+          password: ${{ secrets.DOCKER_PASSWORD }}
+
+      - name: Build Docker image
+        run: docker build -t pyforestscan .
+
+      - name: Tag Docker image
+        run: docker tag pyforestscan iosefa/pyforestscan:latest
+
+      - name: Push Docker image
+        run: docker push iosefa/pyforestscan:latest
