### 2024-12-03 4ef7f71
Add new references and links, update acknowledgments
This commit adds new references in the bibliography and updates citations in the paper's markdown file. It includes precise URLs for Jupyter notebooks and additional acknowledgment for Ryan Perroy's contributions. The .gitignore file is also updated to exclude .pdf and .jats files.
