+++ paper/paper.bib
@@ -366,3 +366,21 @@ abstract = {Obtaining accurate and widespread measurements of the vertical struc
   url          = {https://copc.io/},
   note         = {Latest version accessed on 2024-11-22}
 }
+
+@article{tangAlgorithmTheoreticalBasis2019,
+  title = {Algorithm {{Theoretical Basis Document}} ({{ATBD}}) for {{GEDI L2B Footprint Canopy Cover}} and {{Vertical Profile Metrics}}},
+  author = {Tang, Hao and Armston, John},
+  year = {2019},
+  month = dec,
+  langid = {english},
+  file = {/Users/iosefa/Zotero/storage/IBWZ8BJW/Tang and Armston - GEDI L2B Footprint Canopy Cover and Vertical Profi.pdf}
+}
+
+@misc{NOAA_HI_Lidar_2019,
+  author       = {Office for Coastal Management},
+  year         = {2024},
+  title        = {2018 - 2020 NOAA USGS Lidar: Hawaii, HI from 2019},
+  publisher    = {NOAA National Centers for Environmental Information},
+  url          = {https://www.fisheries.noaa.gov/inport/item/68082},
+  note         = {Accessed: 2024}
+}
