### 2024-12-03 e6e4c01
Update conda channel and visualization function names
Switch conda channel to conda-forge for package installation and rename visualization function to a more general term. These changes ensure compatibility with conda-forge packages and improve code readability by using a generic function name for plotting metrics.
