### 2024-12-03 iosefa/PyForestScan#11 @iosefa
Thank you very much for your review of the paper. I have implemented the suggestions you made, including several more that were requested by the other reviewer. These changes were made to better organize the description of the software. This was done to address the comments by the other reviewer, specifically those related to the issues of the missing “state of the field” and the concern for the lack of substantial scholarly effort. These included moving the introduction of existing software to be earlier in the text, and expanding on this to provide a more detailed picture of the state of the field. These also included reorganizations in the text to enhance emphasis on the functions that this library introduces to calculate forest structural metrics in Python, something that prior to this library did not exist, and reducing the emphasis on the contribution of the library in solving “big data” and efficiency problems, as the other reviewer is right to point out that I am leveraging pdal to do this. Many of the required changes warranted more explanation and so the revision of the manuscript is considerably longer. Finally, I think that it is important to say that this is the first stable release of the library. I hope and plan to continue to add additional functionality, like the ability to calculate canopy cover, incorporate tree segmentation, and work with waveform lidar – for example modeling point cloud distributions from waveforms.

To address the points you raised:

> 1. **Title Consistency:** The title uses "LiDAR," while the main text refers to it as "lidar." Please ensure consistency throughout.

I changed all the text to use `lidar`.

> 2. **Term Capitalization:** In the summary, terms like "foliage height diversity" (FHD) are presented in all lowercase, whereas in the Statement of Need, they are capitalized as "Foliage Height Diversity" (FHD). The same inconsistency applies to PAD and PAI. Please standardize the capitalization across the document.

I standardized capitalization across the document to all be in lowercase.

> 3. **Citation Recommendation:** Consider citing PDAL: https://doi.org/10.5281/zenodo.2616780.

I added this reference and another for PDAL: https://doi.org/10.1016/j.cageo.2020.104680 -- thank you for sharing the zenodo reference!
