### 2024-12-03 ec84d89
Add Python 3.13 to CI test matrix
This update includes Python 3.13 in the continuous integration test matrix. Ensuring compatibility with this version helps future-proof the project and allows us to catch any potential issues early.
