### 2024-12-03 7ebc5a3
Add attribution section for PDAL in documentation
The new section acknowledges the use of PDAL for IO operations in the library, crediting its significant role in simplifying the calculation of forest structural metrics. Citations for the relevant PDAL publications are included to enhance the documentation's scholarly robustness.
