### 2024-12-03 iosefa/PyForestScan#12 @iosefa
Thank you for your detailed review and the excellent suggestions on how to improve the repository. I implemented all of the suggestions and I am continuously working to improve the repository and accessibility of the library. Your comments here were very helpful and I appreciated your suggestion to refer to the segment-geospatial package!

I provided an overview of the changes to the manuscript and the issues addressed in #11 in https://github.com/iosefa/PyForestScan/issues/11#issuecomment-2514117784.

I will now address the major concerns raised in this issue:

> Two critical elements are currently absent:
>
> Automated Tests: There are no automated tests to verify the software's functionality and Python version support (3.10-3.13)
> Community Guidelines: While a CONTRIBUTING.md file exists in the repository, no community guidelines are available on the documentation website.

Automated Tests: We improved automated testing to include Python versions 3.10 - 3.13. Previously, tests were only conducted on Python version 3.10. There can be seen here: https://github.com/iosefa/PyForestScan/blob/ec84d89379929028e067e49162ff964d9c751283/.github/workflows/main.yml#L15-L17

Community Guidelines: We moved the contributing.md file to the docs directory and now both the README in the repository and the documentation have references to it. Thank you!

> Sample Datasets and Jupyter Notebooks: Providing sample datasets along with complete Jupyter Notebook examples would facilitate testing of the package's functionality.
> Binder Integration: Link to [mybinder.org](https://mybinder.org/) to allow users to try out notebooks without local installation. Although installing PDAL on Google Colab can be challenging, using a Dockerfile to set up the environment for Binder could streamline this process.
> Data Visualization Examples: The package includes data visualization features that are not showcased in the Quick Start Example.
> Reference to PDAL: Enhance the references section by citing PDAL (https://doi.org/10.5281/zenodo.2616780), as it is a crucial dependency.

Sample Datasets and Jupyter Notebooks: We took inspiration from the samgeo repository and we moved the jupyter notebooks to the docs directory so that they can also be included in the documentation. We created additional notebooks -- all are in the docs/examples directory. I added more sample data to the the docs/example_data. Thank you for this suggestion.

Binder Integration: We added binder integration by created a Docker image. This image is hosted on Docker Hub and the binder notebook will run the latest docker image.

Data Visualization Examples: We made heavy modifications to the documentation -- including moving the docs from readthedocs to mkdocs. We thought hard on this and felt that if there was a time to do it, it is now! Given the user community is still very small, we felt that we could make some of these larger changes to the repository. The docs are now better organized and include the visualizations that were missing.

Reference to PDAL: This was a very important reference, and it has been added to both the paper and the documentation. Thank you.

> It appears that the submitting author is the sole contributor of source code to this repository. Clarification on the roles of the two co-authors in the development of this package would be beneficial.

We added a Contributions section to the manuscript. iosefa did contribute a significant majority of the source code, but this was mostly done together with benleamon -- either through pair programming or discussion. benleamon presented the first version of the software to FOSS4G Asia and, while he has not contributed as much code, he has had an important impact on the develop of this library. For this reason, I think that it is important that he remains an author. Ryan Perroy provided guidance on the direction of the project during its later stages, and helped in the writing of the manuscript. However, seeing that this may not qualify, we have agreed to remove him as an author and instead include him in the acknowledgements in recognition of his contributions and support.

> Enhance Repository Information: Add a project description, website, and relevant tags to the About section (next to the green Code button).
> Include Badges: Add badges (e.g., PyPI, Binder) to the README.md for better visibility and usability.

Enhance Repository Information: We added a project description, website (the new documentation site), and relevant tags.

Badges: We added badges to both the README and the documentation home page. Thank you for the suggestion!
