### 2024-12-03 2c5f219
Add jupyter notebook examples and installation guide
Updated the documentation to include guidance on using jupyter notebooks with examples to help users get started. Added instructions for installing jupyter using either conda or pip for easy setup.
