### 2024-12-03 #7314 @iosefa
@mengqi-z sorry for the delay and thank you for your patience!

I have responded to all the reviewer comments and have incorporated all of their feedback into the manuscript, the library, and the library documentation.

I consolidated all of my responses in comments to the issues. I hope that this was the correct thing to do. I also tagged the reviewers, and I worry now that maybe that was not the thing to do...

Here are my responses to the reviewer feedback contained within the 4 issues that were created:

- https://github.com/iosefa/PyForestScan/issues/11#issuecomment-2514117784
- https://github.com/iosefa/PyForestScan/issues/12#issuecomment-2514218328
- https://github.com/iosefa/PyForestScan/issues/13#issuecomment-2514275004
- https://github.com/iosefa/PyForestScan/issues/14#issuecomment-2514329605

And I would like to also address with you the comment from Dr. Pucino about his concern for substantial scholarly effort.

The library that we have created and are writing this manuscript for introduces functionality in the Python ecosystem to calculate fundamental and very important metrics of forest structure. These metrics include plant area index, plant area density, foliage height diversity, and canopy height. To my knowledge, there is no existing published python library that offers the tools to calculate these metrics. These metrics are very important in ecology, for not just forests, but also birds and other biological populations, as well as for studies in landcover dynamics and climate. Brining these tools to Python is important because a lot of science is done in Python and this allows the calculation of these metrics to be brought into routines and analysis done in Python. Platforms such as Google Earth Engine have Python APIs that are heavily used in remote sensing, ecology, and land cover studies. Deep learning tools are very active in Python and deep learning tools play a very important role in studies on forest ecology, and so we feel that having the tools to calculate these metrics in Python is an important contribution. However, the point about the dependence on PDAL was well taken. The library uses PDAL primarily for its IO operations and preprocessing steps. These include reading and writing to various point cloud formats, and standard routines to filter and clean point cloud data. However, PDAL does not provide features to calculate metrics like PAI, PAD, and FHD. We have heavily revised the manuscript to place greater emphasis on the structural metrics that this library allows researchers to calculate, and less so on the performance gains that the library provides. As I mention in my response to the reviewers, this is not to say that this isnt an important aspect of the software. Performance is quite important because many of these metrics need to be calculated at very large landscape scales, and so by having the functions that our library introduces, like the ability to calculate PAI, be 20 times faster than existing solutions outside of Python, and this performance boost does not depend on PDAL, research can be conducted much faster. Then, with the additional capabilities of PDAL, which we leverage for IO operations, land cover research to calculate these metrics can then be done that much faster and without the use of dedicated HPC clusters. Most (or all) or the existing solutions outside of Python to calculate some of these metrics do not make use of highly efficient data formats, like EPT, or leverage the capabilities of PDAL to improve their operations. We have updated both our manuscript and the documentation to provide proper citation to PDAL.

I hope that this response helps clarify the work that we have done and what we feel are its contributions. The responses that both reviewers gave were incredibly helpful and have helped make this work a lot stronger.

That all said, the branch `docs/joss-paper` (https://github.com/iosefa/PyForestScan/tree/docs/joss-paper) now has the latest version of the manuscript and is ready for whatever the next step is. Please let me know what I need to do next. Thank you!
