### 2024-12-03 a8d61b2
Fix Python version syntax in workflow configuration
Correct a syntax error in the list of Python versions in the GitHub Actions workflow YAML file. This change ensures that version numbers are properly separated, preventing potential issues during CI/CD pipeline execution.
