### 2024-12-03 df31dcd
Add .dockerignore and update Dockerfile for example data
Introduce a .dockerignore file to exclude unnecessary files like .tif and example data paths from Docker builds. Update the Dockerfile to copy example_data into the image, ensuring all necessary resources are included during the build process.
