### 2024-12-03 iosefa/PyForestScan#20 @mengqi-z
JOSS Paper - Invalid DOIs in the Reference
Hi @iosefa,

I noticed a few missing and invalid DOIs in the latest version of your paper for the JOSS submission https://github.com/openjournals/joss-reviews/issues/7314. You can find the details here: https://github.com/openjournals/joss-reviews/issues/7314#issuecomment-2515213230.

To resolve the issue with the invalid DOIs, please remove the `https://doi.org/` prefix from the DOI links in your `paper.bib` file.

Thanks!
