### 2024-12-03 646b885
Remove redundant 'https://' prefix from DOI entries
This commit updates the DOI entries in several bibliography items by removing the unnecessary 'https://' prefix. The change improves consistency with common bibliographic standards and ensures that DOI links are formatted correctly without redundant URL components.
