### 2024-12-03 iosefa/PyForestScan#20 @iosefa
Hello @mengqi-z

I just made the change to fix the invalid DOIs and pushed to the paper branch. There is one invalid DOI that is actually valid -- the one about leafR:

> ❌ MISSING DOIs
>
> - 10.32614/cran.package.leafr may be a valid DOI for title: leafR: Calculates the Leaf Area Index (LAD) and Ot...

There was this thread about it earlier in the submission process: https://github.com/openjournals/joss-reviews/issues/7283#issuecomment-2375548957

So, I wont change that one as I believe that it is actually correct.

Thank you!
