### 2024-12-03 iosefa/PyForestScan#20 @mengqi-z
Hi @iosefa,

Thanks for revising the DOIs.

Regarding the DOI for `leafR`, it was missing from your Reference, not invalid. The JOSS editorial robot was able to locate the DOI, which matches the one listed on [`leafR`'s CRAN page](https://cran.r-project.org/web/packages/leafR/index.html): `10.32614/CRAN.package.leafR`. If you're using this version of `leafR` (v0.3.5), it would be appropriate to add this DOI in the citation, in addition to its url.

```
@report{dealmeidaLeafRCalculatesLeaf2021,
	title = {{leafR: Calculates the Leaf Area Index (LAD) and Other Related Functions}},
	url = {https://CRAN.R-project.org/package=leafR},
        doi = {10.32614/CRAN.package.leafR},
	type = {manual},
	author = {de Almeida, Danilo Roberti Alves and Stark, Scott Christopher and Silva, Carlos Alberto and Hamamura, Caio and Valbuena, Ruben},
	date = {2021},
}
```

Thanks!
