### 2024-12-03 bfb9803
Add DOI to leafR manual reference in bibliography
Included the DOI for the leafR manual citation to ensure proper referencing and easy access. This enhances the academic integrity and traceability of the cited resource within the paper.
