+++ paper/paper.bib
@@ -59,6 +59,7 @@
 @report{dealmeidaLeafRCalculatesLeaf2021,
 	title = {{leafR: Calculates the Leaf Area Index (LAD) and Other Related Functions}},
 	url = {https://CRAN.R-project.org/package=leafR},
+	doi = {10.32614/CRAN.package.leafR},
 	type = {manual},
 	author = {de Almeida, Danilo Roberti Alves and Stark, Scott Christopher and Silva, Carlos Alberto and Hamamura, Caio and Valbuena, Ruben},
 	date = {2021},
