### 2024-12-09 iosefa/PyForestScan#13 @npucino
Hi @iosefa,
thanks for taking the time to carefully address each individual comment and improve your work. I am impressed by the quality of your revision and I very much like the way you restructured the paper. Now the importance and scholarly effort of this contribution is way clearer, and I am very happy to accept this submission!

Congrats!
with regards,
Nick
