### 2024-12-09 iosefa/PyForestScan#14 @npucino
Hi @iosefa,
thx for taking the time to adress these comments.
I am very happy to see your effort in making clear the importance of your work and I am convinced that now, the usefulness and role of this package in the Pythonic Forestry community.
I still do not understand how SfM can 3D model a whole forest to justify the use of pyforestscan, but, I agree with you that this package will work with point clouds no matter how they are generated so if there are scientists out there crazy enough to model a forest using SfM and really getting vertically distributed tie-points (and not only the top-of-the-canopy), then this package can be used!
Overall, I am happy to accept this submission as it is now and I want to congratulate you all with the initial development and the revision of it.

Congrats!
with regards,
Nick
