### 2024-12-12 c55451b
Update documentation link in README.md
Replaced the outdated documentation URL with the correct one pointing to the new domain. This ensures users are directed to the most up-to-date and accurate resources.
