### 2025-01-13 #7314 @mengqi-z
@giswqs - Thanks for the update!!

@iosefa - Both reviewers have completed their reviews. I’ll proofread the paper and ensure it meets JOSS standards, then follow up with you on the next steps. Thanks!
