### 2025-01-14 iosefa/PyForestScan#22 @mengqi-z
JOSS Paper
Hi @iosefa,

Thanks for your submission to JOSS https://github.com/openjournals/joss-reviews/issues/7314. I have a few comments regarding your paper that need to be addressed before we proceed to the next steps:

1. Reference for Roussel et al., 2024: A DOI is available for this reference (10.32614/CRAN.package.lidR). Please use the DOI instead of the page url.

2. Author Name Correction: In the reference "Coastal Management, O. for. (2024).", please update the author name so that it shows as "Office for Coastal Management" instead of "Coastal Management, O. for."

3. Line 33: The terms FHD and PAI are already defined in the summary, so there’s no need to redefine them here.

4. Package Name Formatting: Package names such as PyForestScan, lidR, canopyLazR, and others should be formatted as code style using backticks (e.g., \`PyForestScan`). Please review the entire paper and ensure all package names are updated accordingly.
