### 2025-01-16 3246e54
Address comments for JOSS submission (#22)
