+++ paper/paper.bib
@@ -68,6 +68,7 @@
 @report{rousselAirborneLiDARData2024,
 	title = {Airborne {LiDAR} Data Manipulation and Visualization for Forestry Applications},
 	url = {https://cran.r-project.org/package=lidR},
+	doi = {10.32614/CRAN.package.lidR},
 	type = {manual},
 	author = {Roussel, Jean-Romain and Auty, David},
 	date = {2024},
