### 2025-01-20 94763f6
Fix braces around author field in bibliography entry for Office for Coastal Management
Added extra braces around the author name in the NOAA_HI_Lidar_2019 entry to ensure proper formatting in BibTeX. This change prevents potential issues with capitalization or rendering in citations.
