+++ paper/paper.bib
@@ -379,7 +379,7 @@ abstract = {Obtaining accurate and widespread measurements of the vertical struc
 }
 
 @misc{NOAA_HI_Lidar_2019,
-  author       = {Office for Coastal Management},
+  author       = {{Office for Coastal Management}},
   year         = {2024},
   title        = {2018 - 2020 NOAA USGS Lidar: Hawaii, HI from 2019},
   publisher    = {NOAA National Centers for Environmental Information},
