### 2025-01-21 #7314 @iosefa
@mengqi-z great. Thank you very much! I have copied and pasted the list here so that I can track progress. I will ping you again once everything is done.

- [ ] Double check authors and affiliations (including ORCIDs).
- [ ] Archive the released software on Zenodo or a similar service (e.g., figshare, an institutional repository).
- [ ] Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.
- [ ] Please post the DOI of the archived version here.
