### 2025-01-21 b00c56c
Update version to 0.2.2 in setup.py
Bump the package version from 0.2.1 to 0.2.2 to reflect recent changes. This ensures the package metadata stays in sync with the latest updates or fixes.
