### 2025-01-21 #7314 @iosefa
> [@mengqi-z](https://github.com/mengqi-z) great. Thank you very much! I have copied and pasted the list here so that I can track progress. I will ping you again once everything is done.
>
> * [x]  Double check authors and affiliations (including ORCIDs).* [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
>
>     * Release version is `v.0.2.2`[x]  Archive the released software on Zenodo or a similar service (e.g., figshare, an institutional repository).[x]  Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.[x]  Make sure that the license listed for the archive is the same as the software license.* [x]  Please post the DOI of the archived version here.
>
>     * DOI: https://doi.org/10.5281/zenodo.14714890

@mengqi-z done!
