### 2025-01-21 #7314 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
