### 2025-01-27 #7314 @kthyng
@iosefa The paper is pretty long — I see it gained words since the pre-review. Can some of this material move to the docs or readme?
