### 2025-01-27 #7314 @iosefa
> [@iosefa](https://github.com/iosefa) The paper is pretty long — I see it gained words since the pre-review. Can some of this material move to the docs or readme?

Hi @kthyng - Yes, we added additional text in response to some of the reviewer comments. However, I can revise it further and move some of the content, probably the math section, to the docs. Shall I go ahead with this and try to make it under the 1000-word normal limit?
