### 2025-02-02 4dba083
Simplify equations section in `PyForestScan` documentation
Removed inline equations for PAD, PAI, and FHD in the manuscript and referenced the documentation for detailed mathematical descriptions. This streamlines the text, making it more concise while still providing access to necessary information.
