### 2025-02-03 4fca117
Simplify phrasing and update tone in paper.md
Refined wording and adjusted tone for clarity and conciseness in the paper's "Summary," "Statement of Need," and "PyForestScan" sections. These edits enhance readability and ensure consistency while maintaining the technical accuracy of the content.
