### 2025-02-03 03507c2
Refine description of `PyForestScan` capabilities.
Clarified the limitations of existing tools and emphasized `PyForestScan`'s unique features, such as native tiling, multi-format support, and its focus on scalable forest analysis. These updates align the narrative with the tool's core strengths and application in large-scale ecological research.
