### 2025-02-03 #7314 @iosefa
Hi @kthyng - ready for your review again! I removed content from the paper that exists in the documentation and added a note to refer readers to the docs. I also made minor changes to make some of the text a little more concise. You can see the changes here:
- https://github.com/iosefa/PyForestScan/commit/4dba083af07296d0054841812d406b8abbafdf8a
- https://github.com/iosefa/PyForestScan/commit/c58119c27e9d554eccf81738bf6a1b20b97f8393
- https://github.com/iosefa/PyForestScan/commit/4fca1171cfeb29691c4cbe0380105021f1cfa38a
- https://github.com/iosefa/PyForestScan/commit/03507c2ce3b9d29d057df077077d09a4c8d3fde4
