### 2025-02-06 10934ec
Format bibliographic titles using curly braces for consistency.
Enclosed titles in curly braces in the BibTeX file to ensure proper formatting and capitalization when processed by reference management tools. This improves the appearance of titles in generated bibliographies.
