+++ paper/paper.bib
@@ -297,7 +297,7 @@ Use of these data requires citation of this dataset},
 
 @software{manning_entwine,
   author       = {Connor Manning},
-  title        = {Entwine: Open Source Point Cloud Indexing},
+  title        = {{Entwine: Open Source Point Cloud Indexing}},
   year         = {2024},
   url          = {https://entwine.io/},
   note         = {Latest version accessed on 2024-11-22}
@@ -348,7 +348,7 @@ abstract = {As large point cloud datasets become ubiquitous in the Earth science
 }
 
 @article{DUBAYAH2020100002,
-title = {The Global Ecosystem Dynamics Investigation: High-resolution laser ranging of the Earth’s forests and topography},
+title = {{The Global Ecosystem Dynamics Investigation: High-resolution laser ranging of the Earth’s forests and topography}},
 journal = {Science of Remote Sensing},
 volume = {1},
 pages = {100002},
