### 2025-02-06 #7314 @iosefa
> [@iosefa](https://github.com/iosefa) paper length looks good! Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, "mediterranean" is there but please check all references.

@kthyng Thank you. I just finished reviewing the reference list. The original paper by Guerra-Hernandez et al. did not capitalize "Mediterranean", so I am not sure if I should. See https://doi.org/10.1080/22797254.2024.2315413

There were some titles, though, that I did need to change by adding the curly brackets. The changes are: [10934ecb30f4dab361db0c1204a766cbae0ed275](https://github.com/iosefa/PyForestScan/commit/10934ecb30f4dab361db0c1204a766cbae0ed275)
