### 2025-02-08 #7314 @iosefa
> Congratulations on your new publication [@iosefa](https://github.com/iosefa)! Many thanks to editor [@mengqi-z](https://github.com/mengqi-z) and to reviewers [@npucino](https://github.com/npucino) and [@giswqs](https://github.com/giswqs) for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
>
> Note we have a new tool for reviewers! You can go to https://joss.theoj.org/papers/reviewed_by/@your-github-username to see the JOSS submissions you have reviewed, and you can also copy a badge there with the number of your JOSS reviews.
>
> [@iosefa](https://github.com/iosefa) If you'd like to join JOSS as a reviewer, please sign up at https://reviewers.joss.theoj.org/join!

@kthyng Thank you very much! I just signed up to be a JOSS reviewer.

Best regards,
sefa
