### 2025-02-09 #7314 @npucino
Thx, it was a pleasure to contribute to JOSS!
However, I checked the new reviewers tool but it does not find my reviews. I reviewed 2 papers if I am not mistaken including this one but nothing shows up.

CHeers

Nicolas Pucino (He/Him)

Postdoctoral Researcher @ Australian National University | Fenner School of Environment and Society

Visiting Scientist @ CSIRO Environment | Living Landscapes Program, Earth Observation

ISPRS SC Vice President 2022-2026 | International Society of Photogrammetry and Remote Sensing

[cid:93fd26ae-5c00-4cf3-bcb7-c13a4f0aad25]     [cid:5daf2325-8923-4536-9911-abcdc34cf1ee]      [cid:709bf620-085d-4d18-8dca-94b4d0360887]
