### 2025-02-10 #7314 @kthyng
I see the same thing as @mengqi-z. To see if I could find other reviews you've done, I also checked here: https://github.com/openjournals/joss-reviews/issues?q=is%3Aissue%20npucino and I see your handle pop up in a bunch of issues but for other reasons, not as a reviewer. Please let us know if you can find a link and if so we can pass it to dev.
